+++ packages/Label/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.1.10
+
+### Patch Changes
+
+- [`f23eea3`](https://github.com/electronic33/ag-ui-react/commit/f23eea3ad84886203be361f5c781cb97237b19c0) Thanks [@Levi1235](https://github.com/Levi1235)! - rebuild
+
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.