+++ packages/footer/CHANGELOG.md
@@ -1,15 +1,26 @@
 # Change Log
 
+## 1.0.4
+
+### Patch Changes
+
+- [`9e7f2c1`](https://github.com/electronic33/ag-ui-react/commit/9e7f2c138999ec18e4c8e9196ba528214d9243ab) Thanks [@Levi1235](https://github.com/Levi1235)! - removed react icons
+
+## 1.0.3
+
+### Patch Changes
+
+- [`f23eea3`](https://github.com/electronic33/ag-ui-react/commit/f23eea3ad84886203be361f5c781cb97237b19c0) Thanks [@Levi1235](https://github.com/Levi1235)! - rebuild
+
+- Updated dependencies [[`f23eea3`](https://github.com/electronic33/ag-ui-react/commit/f23eea3ad84886203be361f5c781cb97237b19c0)]:
+  - @app-garage/label@1.1.10
+
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 ## 1.0.2 (2021-03-24)
 
 **Note:** Version bump only for package @app-garage/footer
-
-
-
-
 
 ## 1.0.1 (2021-03-21)