+++ packages/numeric-input/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 1.0.6
+
+### Patch Changes
+
+- [`9e7f2c1`](https://github.com/electronic33/ag-ui-react/commit/9e7f2c138999ec18e4c8e9196ba528214d9243ab) Thanks [@Levi1235](https://github.com/Levi1235)! - removed react icons
+
+## 1.0.5
+
+### Patch Changes
+
+- [`3943150`](https://github.com/electronic33/ag-ui-react/commit/394315007a549e1bebced2ccd76cbee466e1bbe9) Thanks [@Levi1235](https://github.com/Levi1235)! - fixed packages
+
+## 1.0.4
+
+### Patch Changes
+
+- [`f23eea3`](https://github.com/electronic33/ag-ui-react/commit/f23eea3ad84886203be361f5c781cb97237b19c0) Thanks [@Levi1235](https://github.com/Levi1235)! - rebuild
+
+- Updated dependencies [[`f23eea3`](https://github.com/electronic33/ag-ui-react/commit/f23eea3ad84886203be361f5c781cb97237b19c0)]:
+  - @app-garage/label@1.1.10
+  - @app-garage/button@1.1.7
+
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.