+++ packages/slider/CHANGELOG.md
@@ -1,15 +1,38 @@
 # Change Log
 
+## 1.0.5
+
+### Patch Changes
+
+- [`9e7f2c1`](https://github.com/electronic33/ag-ui-react/commit/9e7f2c138999ec18e4c8e9196ba528214d9243ab) Thanks [@Levi1235](https://github.com/Levi1235)! - removed react icons
+
+- Updated dependencies [[`9e7f2c1`](https://github.com/electronic33/ag-ui-react/commit/9e7f2c138999ec18e4c8e9196ba528214d9243ab)]:
+  - @app-garage/modal@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- [`3943150`](https://github.com/electronic33/ag-ui-react/commit/394315007a549e1bebced2ccd76cbee466e1bbe9) Thanks [@Levi1235](https://github.com/Levi1235)! - fixed packages
+
+- Updated dependencies [[`3943150`](https://github.com/electronic33/ag-ui-react/commit/394315007a549e1bebced2ccd76cbee466e1bbe9)]:
+  - @app-garage/modal@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- [`f23eea3`](https://github.com/electronic33/ag-ui-react/commit/f23eea3ad84886203be361f5c781cb97237b19c0) Thanks [@Levi1235](https://github.com/Levi1235)! - rebuild
+
+- Updated dependencies [[`f23eea3`](https://github.com/electronic33/ag-ui-react/commit/f23eea3ad84886203be361f5c781cb97237b19c0)]:
+  - @app-garage/modal@1.0.3
+
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 ## 1.0.2 (2021-03-24)
 
 **Note:** Version bump only for package @app-garage/slider
-
-
-
-
 
 ## 1.0.1 (2021-03-21)